Change from unsigned long to uint32_t -- possible 64-bit bug
git-svn-id: https://svn.atnf.csiro.au/difx@6088 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ mpifxcorr/branches/bonnbug/src/nativemk5.cpp
@@ -28,6 +28,7 @@
 #include <mpi.h>
 #include <unistd.h>
 #include <vdifio.h>
+#include <stdint.h>
 #include "config.h"
 #include "nativemk5.h"
 #include "watchdog.h"
@@ -648,7 +649,7 @@ int NativeMk5DataStream::readonedemux(bool resetreference, int buffersegment)
   }
 
   //cinfo << startl << "At the beginning of readonedemux, readpointer is " << readpointer << endl;
-  rbytes = moduleRead((unsigned long*)datamuxer->getCurrentDemuxBuffer(), datamuxer->getSegmentBytes(), readpointer, buffersegment);
+  rbytes = moduleRead((uint32_t *)datamuxer->getCurrentDemuxBuffer(), datamuxer->getSegmentBytes(), readpointer, buffersegment);
   //the main readpointer will be updated outside of this routine, use localreadpointer for here
   localreadpointer = readpointer + rbytes;
   if(rbytes != datamuxer->getSegmentBytes()) {
@@ -658,7 +659,7 @@ int NativeMk5DataStream::readonedemux(bool resetreference, int buffersegment)
   while(fixbytes > 0) {
     ++nfix;
     readto = reinterpret_cast<char*>(datamuxer->getCurrentDemuxBuffer()) + rbytes - fixbytes;
-    moduleRead(reinterpret_cast<unsigned long*>(readto), fixbytes, localreadpointer, buffersegment);
+    moduleRead(reinterpret_cast<uint32_t *>(readto), fixbytes, localreadpointer, buffersegment);
     readpointer += fixbytes; //but if we need extra reads, then we must add these "extra" values to the readpointer
     localreadpointer += fixbytes; //and to the local one also of course
     fixbytes = datamuxer->datacheck(datamuxer->getCurrentDemuxBuffer(), rbytes, rbytes - fixbytes);
@@ -675,7 +676,7 @@ int NativeMk5DataStream::readonedemux(bool resetreference, int buffersegment)
   return rbytes;
 }
 
-int NativeMk5DataStream::moduleRead(unsigned long *destination, int nbytes, long long start, int buffersegment)
+int NativeMk5DataStream::moduleRead(uint32_t *destination, int nbytes, long long start, int buffersegment)
 {
 	int bytes = nbytes;
 	XLR_RETURN_CODE xlrRC;
@@ -760,7 +761,7 @@ int NativeMk5DataStream::moduleRead(unsigned long *destination, int nbytes, long
 
 void NativeMk5DataStream::moduleToMemory(int buffersegment)
 {
-	unsigned long *buf, *data;
+	uint32_t *buf, *data;
 	char *readto;
 	int rbytes, obytes;
 	double tv_us;
@@ -774,7 +775,7 @@ void NativeMk5DataStream::moduleToMemory(int buffersegment)
 	int n = 0;
 
 	/* All reads of a module must be 64 bit aligned */
-	data = buf = reinterpret_cast<unsigned long *>(&databuffer[buffersegment*(bufferbytes/numdatasegments)]);
+	data = buf = reinterpret_cast<uint32_t *>(&databuffer[buffersegment*(bufferbytes/numdatasegments)]);
 
 	waitForBuffer(buffersegment);
